Implements most (all?) of Mol* contacts and creates a (hopefully final?) contacts and interactions interface. Another big refactor is done with the atom typing code, which now uses user literals and constexpr to simplify code and make it readable. There is also a lot of thought and work that has gone into making these literal work with compile-time analysis and static type checking. The order of the AtomType enum is also guaranteed to never change with static type checking picking up any re-ordering of the values.
I am now also using FastNS for neighbor searching (ahead of brute forcing or KDTree). This
is extremely fast (< 1ms), but it won't work for many cases yet, because of cutoff constraints
(this will be fixed in the next commit).

I have also added AtomData (AtomDataVec) to represent atom-level information, Feature and
FeatureVec to represent group-level data, and finally, I am in the process of refactoring the
ring-detection and storage interface. For now, I have updated it to store atom pointers (instead
of atom ids), and simplified a few things.