Improved the ring and residue interfaces.
Rings are now much more correctly identified using size-specific depth-first-search.
This is intended to work with predefined residues that are known to be aromatic, but allows
for a much more dynamic identification of aromatic atoms within the residue (as opposed
to predefined atom types). I also now handle RDKit RingInfo by manually initializing and
populating it with identified rings. The interface to split the molecule into known and
unknown parts has also been improved. This too, will correctly populate RDKit's RingInfo.

Rings are added to the group-typing strategies and this allowed for the simplification
of the molstar-based typing mechanism. One downside of this is that group-type identification
is required to use ring information, as it is now its responsibility to correctly initialize
and populate RingInfo. The current interface hides this, and leads to confusion behavior,
such as making the order of atom-type and group-type initialization dependent on the order.
(group-type -> atom-type works, but the reverse does not).

Further improvement is made by adding a Residue-level abstraction. The implementations of
group-typing strategies is clearer now. I also added a convenience holder of Feature's
that's the return type of these strategies. This makes it easier and clearer when using
center coordinates, for instance.