Created a centralized interface to compute neighbors
The issue with neighbor searching for interactions was that in many
cases the filtered atom types were too few and the cutoff-criteria for
grid-based searching were not met. The current solution using a simple
heuristic to determine if grid-based or a simple brute force should be
used. Brute force will also be used if, for any reason, grid-based is
not appliable.

Cleaned up the different interactions to use this new implementation.