Benchmarked scaling performance of the recent changes.
Specifically, made sure the many changes introduced with the new
interaction interface did not brick the scaling of the software. The
only issue was with `in_aromatic_ring_with_N_or_O` which was making a
copy of the ring data, instead of taking a const reference. I also added
an early return if the atom is not aromatic.