Refactored entity representation (Atom, Ring, and Group entities)
Migrated the entity-related code to two files to represent a single
Entity object (either AtomEntity, RingEntity or GroupEntity) and a
collection of entities (e.g. AtomEntityCollection). The representation
for each of these categories (single object and vector or objects) has
been rewritten to highlight their shared functionality.

This significantly improves the code quality and readability, not to
mention making it much easier to maintain in the future. The API for
getting these entites is also now consistent in its implementation.

One important thing to note is that I decided to rely on type-erasure
through function pointers for the implementation and mainly to avoid
virtual functions which I perceive (though I have not tested) to affect
performance. Still, this implementation is not without overhead and
that's something that I need to test in the future. Using std::variant
could also provide better safety and avoid using `const void*`.