Got rid of KDTree-based search support
Based on a few benchmark tests, the existing KDTree implementation was
not providing any performance benefits. I, therefore, decided to remove
it to avoid unnecessary bloat. I also got rid of the old code that
heuristically chose KDTree vs. the brute force approach and was
prototyped to work with group type entities.