Improved contacts and handling of bond edge cases
Added a function to check for duplicates in the computed contacts. This
can happen if two or more atoms have the required AtomType's, then we
would add the same contact but through using the opposite AtomType from
each atom. It is not yet implemented in all contacts.

Re-checked the accuracy of contacts and added several checks to make
sure that future changes do not break things.

Improved calling and using the valence model.