Implements a starting interface to compute alignment
Adds an interface to process query and target files. Can also handle
directories recursively. I developed a processor class that manages
files, splitting them into chunks and correctly pairing query and target
chunks, both in a query-target context, as well as the case when query
and target is the same. Chunking is entirely configurable and should be
fairly lightweight. The current implementation first collects all files
in one place, which may not always be desirable, but presents a good
starting point.

Developed an Aligner class to manage the alignment of sequences (along
with everything that goes into it: extracting coordinates, prefiltering,
and the actual alignment). Coordinate extraction and alignment have a
multithreaded implementation. This is fairly efficient and quite fast
based on a several test runs. The memory footprint is also quite
manageable.

Both the processor and the aligner are extendible. Processor takes in a
Runner type object, and therefore could be used for other usecases in
the future. The aligner, too, takes in a computer with a clear signature
which should make custom processing of structures quite modular.

Refactored the DirectoryHandler to use a more sophisticated Iterator. It
also defines a `next_chunk` member function to process files in
customizable sizes.