Miscellaneous: New Features, Refactors, Bindings, API, Typing (Part 3)
Improved the definition of predefined aromatic residues. Instead of
using std::unordred_map which was not constexpr friendly and could have
led to issues in the future, we now use a fixed-size array.

Refactored (improve, clean up) residue-handling logic. Old templated
functions are replaced with simple, easy to understand, functions/member
functions. This refactor makes it easy to do fairly complex
selection/filtering of residues.

Building residues is also more idiomatic and, importantly, not done
during initialization. Next, we'll also improve the initialization of
the topology.