Miscellaneous: New Features, Refactors, Bindings, API, Typing (Part 4)
Makes topology construction configurable and consistent with other
classes in how it is constructed. Moved the actual topology-related
computations away from the Luni class to the Topology class. It also
handles errors better.

Made the entity type system more explicit and now instead of returning a
vector of AtomType's, we return AtomEntityCollection instead.

Topology construction, if it fails, does not crash the program. It can,
therefore, be made entirely optional. Computing neighbors is not done in
the topology class and they are not stored.