Property system, Loggs, and Type-Safe Dynamic Data Access
This commit introduces many changes:
A new design to manage data/property access on the main classes. This
will further simplify the `Luni` class, and, more importantly, will
enable dynamic as well as variable access to getter member functions in
a type safe manner. This is intended to "extract" data when processing
many files and store them in a thread-safe manner. The current
implementation also adds Python bindings and makes sure that memory is
managed correctly. There is a very small overhead and calling indirection
that this causes but it should be negligable (confirmed from a few test
runs). The system is generic enough that we can use it for other classes
as well.

Registered a few getter member functions. Will need to add the others
later. Adding new properties is very straightforward.

A centralized Logging utility. Updated all the places where we were
using spdlog to now use this Logger class.