Configurable memory threshold
We have an option to avoid exploding memory usage when pooling results
in memory. A default value of 1GB is set as the max threshod before no
longer results will be stored. This is configurable but the API is not
yet accessible via Python.

I also added templated member functions to property results to get the
size of the stored data.

Caveats:
- Most of the memory usages are estimates as it is difficult to be exact
  (and largely irrelevant)
- For RingInfo, I had to change the source files to provide getters to
  the underlying private data.
- Storing bonds and atoms takes up the most space, but is largely
  reasonable. It is certainly a very challenging part of the application
  to optimize.