Adds bindings for alignment and mapping code
The functionality to do "topological equivalency" has already existed
for a while, but now it is also possible to use it from Python. The
interface, certainly, has a few rough edges but it is functional.

As part of this, all configurable options and processing code can be
used in Python. Stub files are also written (though not formatted and
likely to change). It's an open question if the user-interaction should
be directly with the bound code or with another thin pure Python layer.
There really wouldn't be any noteworthy performance impact.

I also removed the stl ctpl usages, and replaced them with only the main
one, which was causing very weird crashing issues.