Faster handling of implicit hydrogens and other atom data
Instead of using the underlying RDProps to store properties which is
slow for our usecases, I instead added data members to RDKit atom along
with getters and setters. This makes it much more efficient to retrieve
these objects. Of course, the performance-flexibility tradeoffs applies.

I also moved the responsibility to apply the valence model to the
topology class, where it makes more sense.