Remove RDKit::RDProp inheritance from RDKit::Atom
Hopefully, I dont' regret this. I suspect RDProps was causing weird
issues and lagging, and although I have not strictly bencharked it, I suspect it was
causing slowdowns on many heap allocations.

I did not remove it entirely as it is critical to RDKit, but I made it a
pointer data member and used a getter function to access it. This made
refactoring very easy. The basic functionality seems to be preserved
based on a very small number of tests I did. We do not use it in Lahuta,
so from our end it should be fine.