Database Implementation and Read/Write Support
Adds preliminary support for an optional LMDB database. Topology
building is the bottleneck and parsing coordinates does not make much of
a difference in practice. However, this is only true for files that
can be memory mapped. Compressed files need to be read and decompressed
normally. In such cases the database would help.

Having a database built-in may allow future optimizations to be
performed, not to mention offloading many other pre-computations to it.

It is also possible to do a zero-copy database read, and I tested this
but since building the topology is the biggest bottleneck, the overall
performance gains are modest.

Despite the caveats, the database is very lightweight and quite fast, so
it should be useful in a large number of different contexts.