Refactor molecule reading and RDKit molecule creation
We've had an unnecessary steps where we created a gemmi::Structure
object and then used this to create our RDKit molecule. This commit
removes this step by directly creating the RDKit molecule.

We also only process the information that we need. For now only the
atom list (but we'll add others as needed).