Performant Pipeline System for Generic Computational Tasks
Implemented a generic, compile-time optimized, computational pipeline
system. It operates on generic sources to process inputs (e.g. files
within a directory, file list, database keys, in-memory vector, etc.).
The input is fed into a series of stages (e.g. computational task
objects) and then passed to "collector" objects (keep in-memory,
database output, json, etc.).

The pipeline is meant to be easily extendible and allow for future
addition of other tasks, as well as make composability possible. It
ended up being more complex than I anticipated. Largely this is due to
our C++17 compatibility and the necessity to rely on SFINAE.

Low-level objects define CTAD to make usage easier, but I also implemented
a DSL wrapper with operator| overloads and chaining operations to build
the pipeline using a simpler syntax. There is also a pipeline builder
which makes construction easier, but I have not tested this extensively
yet.

The collector interface is aimed to provide back-pressure when writing
results (e.g. to disk), which may lead to result accumulation, heavy
memory usage, and severly affect scaling. It's not the most efficient
design, but from the testing I've done it works quite well.

The overall objective is for the pipeline system to replace the
individual executables into one single binary, simplifying usage
considerably.