Refactored our Entity System & Contact Analysis
Our entity system was still left in its early stages of development.
While functional, and serviceable, it did not really fit with the new
design of the library. In this commit, I have rethought and
re-implemented it. Now it's much cleaner and clearer. Contacts follow
directly from the definition of entities.

I also re-implemented how we search for neighbors among entities. I
tested a "view-based" approach, with lightweight iterators that do not
materialize the objects. However, a cache-friendly approach that just
uses indices is better.