Refactored model building topology perception
I refactored the topological perception for (protein) models so that it
adheres to the same pattern as the generic one.

I also updated the Conformer and parser to better handle the
coordinates. We cannot entirely avoid one full copy of the data due to a
float->double conversion. This may be something we revist at some point
in the future.

The CLI now should be able to store the data in a database and compute
contacts from it. More work needs to be done to update the usage
information.