Support for Python Processing Backend
This commit aggregates changes that together allow for the execution
of Python-heavy tasks through the C++ pipeline using a multiprocessing
backend.

This avoids the Pipeline run from becoming serial, but the performance
is not comparable to the threading backend. Without any Python tasks,
the overhead is negligible and the pipeline run is indeed comparable
between the two backends.

The processing backend supports module-based execution and
notebook-style execution. It does not support <stdin> type of runs.
We now add cloudpickle to the dependency list.

I've also added several new tests to exercise the backends, their
serialization type, and execution modes. As part of this, I fixed
several flaky tests (either because of pytest xdist, or Python C-API
lifetimes, or genuine race condition).