Made the main Luni and Topology classes properly thread safe.
This change was triggered by TSan when run on MD data which
require sharing state across threads, but we were not guarding
against mutations. Technically, any mutation we'd do would be
confined to the first frame, but architecting this is challenging so
that it does not impose any hidden implementation details.

It's easier to just make mutations to the core classes thread safe.
There are no TSan errors on the platform I'm testing on.